cambios menores

+++ LibreriaTD/CAD/CocheCAD.cs
@@ -37,7 +37,7 @@ namespace LibreriaTD.CAD
         /* Modificaremos los kilometros de un coche cuando este coche sea alquilado por 
          * algun cliente y tenga mas de los kilometros que tenia antes de ser alquilado
          */
-        public bool modKM(float modKm)
+        public bool modKM(double modKm)
         {
             bool modifi = false;
             return modifi;
@@ -54,7 +54,7 @@ namespace LibreriaTD.CAD
 
         /*Modificar el precio del coche
          */
-        public bool modPrecio(float mPre)
+        public bool modPrecio(double mPre)
         {
             bool modifi = false;
             return modifi;
